refactor: move and improve UploadFromPresignedURL component and update routing
